[AFEC] Regenerated demo for afec_afec_user_sequence_sam_v71_xult
MH3-67770

+++ apps/afec/afec_user_sequence/firmware/src/config/sam_v71_xult/peripheral/tc/plib_tc1.c
@@ -112,7 +112,7 @@ void TC1_CH0_ComparePeriodSet (uint16_t period)
 /* Read the period value */
 uint16_t TC1_CH0_ComparePeriodGet (void)
 {
-    return TC1_REGS->TC_CHANNEL[0].TC_RC;
+    return (uint16_t)TC1_REGS->TC_CHANNEL[0].TC_RC;
 }
 
 /* Set the compare A value */
